+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Better syntax for navigate and update:
+  > bt navigate b.1.862 "https://google.com"
+  > bt update -tabId b.1.862 -url="http://www.google.com" 
+
 1.4.0 (2022-05-29)
 
 * Added "bt navigate" and "bt update" commands